Move Bazel library classes to a separate bazellib directory.
This is to create better separation between the Bazel components that need to run as part of the VSCode extension and the lower level Bazel components that perform the Bazel functions.

The plan is to eventually separate out the Bazel library calls into a node module. This work prepares for that move.
