+++ src/bazel/bazel_quickpick.ts
@@ -13,10 +13,11 @@
 // limitations under the License.
 
 import * as vscode from "vscode";
+
 import { getDefaultBazelExecutablePath } from "../extension/configuration";
-import { IBazelCommandAdapter, IBazelCommandOptions } from "./bazel_command";
-import { BazelQuery } from "./bazel_query";
 import { BazelWorkspaceInfo } from "./bazel_workspace_info";
+import { IBazelCommandAdapter, IBazelCommandOptions } from "./bazellib";
+import { BazelQuery } from "./bazellib/bazel_query";
 
 /**
  * Represents a Bazel target in a QuickPick items window. Implements the
