+++ src/codelens/bazel_build_code_lens_provider.ts
@@ -14,11 +14,8 @@
 
 import * as vscode from "vscode";
 
-import {
-  BazelWorkspaceInfo,
-  getTargetsForBuildFile,
-  QueryLocation,
-} from "../bazel";
+import { BazelWorkspaceInfo, QueryLocation } from "../bazel";
+import { getTargetsForBuildFile } from "../bazel/bazellib";
 import { getDefaultBazelExecutablePath } from "../extension/configuration";
 import { blaze_query } from "../protos";
 import { CodeLensCommandAdapter } from "./code_lens_command_adapter";
