+++ src/codelens/code_lens_command_adapter.ts
@@ -12,11 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import {
-  BazelWorkspaceInfo,
-  IBazelCommandAdapter,
-  IBazelCommandOptions,
-} from "../bazel";
+import { BazelWorkspaceInfo } from "../bazel";
+import { IBazelCommandAdapter, IBazelCommandOptions } from "../bazel/bazellib";
 
 /**
  * Command adapter to pass arguments to Bazel commands.
